backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic==2.7.1
python-multipart==0.0.9
redis==5.0.4
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
